[level 2] Title: N개의 최소공배수, Time: 0.01 ms, Memory: 4.21 MB -BaekjoonHub

+++ 프로그래머스/2/12953. N개의 최소공배수/N개의 최소공배수.cpp
@@ -2,14 +2,14 @@
 #include <vector>
 
 using namespace std;
-long long gcd(long long a, long long b){
+int gcd(int a,int b){
     if(b==0){
         return a;
     }
-    return gcd(b, a%b);
+    return gcd(b,a%b);
 }
 int solution(vector<int> arr) {
-    long long answer=arr[0];
+    int answer = arr[0];
     for(int i=1;i<arr.size();i++){
         answer=(answer*arr[i])/gcd(answer,arr[i]);
     }
